refactor(app): InputParser를 상속하도록 구현
